Command-line selection of algo and camera + typing

+++ cv/camera/Realsense.py
@@ -35,8 +35,8 @@ class RealsenseCamera(CameraBase):
 
         return frame
 
-    def frame_width(self) -> int:
+    def frame_width(self) -> float:
         return self.width
 
-    def frame_height(self) -> int:
+    def frame_height(self) -> float:
         return self.height
